move contexts to contexts folder
