+++ components/Dashboard/WalletConnection.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from "react";
 import { ConnectWallet } from "../ConnectWallet";
-import { useWallet } from "../WalletContext";
+import { useWallet } from "../../contexts/WalletContext";
 import { ClipboardDocumentIcon } from "@heroicons/react/24/outline";
 export function WalletConnection() {
   const { isConnected, address, balance } = useWallet();
