+++ contexts/AuthContext.tsx
@@ -1,0 +1,33 @@
+'use client';
+
+import { createContext, useContext } from 'react';
+import type { ReactNode } from 'react';
+import type { Session } from 'next-auth';
+import { useSession, signIn, signOut } from "@/utils/auth";
+
+interface AuthContextType {
+  session: Session | null;
+  status: 'loading' | 'authenticated' | 'unauthenticated';
+  signIn: typeof signIn;
+  signOut: typeof signOut;
+}
+
+const AuthContext = createContext<AuthContextType | undefined>(undefined);
+
+export const AuthProvider = ({ children }: { children: ReactNode }) => {
+  const { data: session, status } = useSession();
+
+  return (
+    <AuthContext.Provider value={{ session, status, signIn, signOut }}>
+      {children}
+    </AuthContext.Provider>
+  );
+};
+
+export const useAuth = () => {
+  const context = useContext(AuthContext);
+  if (!context) {
+    throw new Error('useAuth must be used within an AuthProvider');
+  }
+  return context;
+};
